諸々
・不要なコメント削除
・エントリポイントと処理のクラスで分割
